+++ src/OrchardCore.Cms.Web/appsettings.json
@@ -278,8 +278,8 @@
     //}
     //"OrchardCore_ReverseProxy": {
     //  "ForwardedHeaders": "None",
-    //  "KnownNetworks": ["192.168.1.100", "192.168.1.101"],
-    //  "KnownProxies": ["192.168.1.200", "192.168.1.201"],
+    //  "KnownNetworks": ["192.168.1.0/24", "192.168.2.0/24"],
+    //  "KnownProxies": ["192.168.1.200", "192.168.1.201"]
     //},
     //"OrchardCore_Facebook": {
     //  "AppId": "",
