+++ src/OrchardCore.Modules/OrchardCore.ReverseProxy/Configuration/ReverseProxyConfigurationMarker.cs
@@ -1,0 +1,6 @@
+namespace OrchardCore.ReverseProxy.Configuration;
+
+/// <summary>
+/// Marker service to indicate that reverse proxy configuration is being managed via appsettings.json.
+/// </summary>
+internal sealed class ReverseProxyConfigurationMarker;
