+++ src/OrchardCore.Modules/OrchardCore.ReverseProxy/Startup.cs
@@ -6,7 +6,7 @@ using OrchardCore.DisplayManagement.Handlers;
 using OrchardCore.Modules;
 using OrchardCore.Navigation;
 using OrchardCore.ReverseProxy.Drivers;
-using OrchardCore.ReverseProxy.Services;
+using OrchardCore.ReverseProxy.Configuration;
 using OrchardCore.ReverseProxy.Settings;
 using OrchardCore.Security.Permissions;
 using OrchardCore.Settings.Deployment;
@@ -29,10 +29,7 @@ public sealed class Startup : StartupBase
         services.AddPermissionProvider<Permissions>();
         services.AddSiteDisplayDriver<ReverseProxySettingsDisplayDriver>();
 
-        services.AddSingleton<IReverseProxyService, ReverseProxyService>();
-
         services.AddTransient<IConfigureOptions<ReverseProxySettings>, ReverseProxySettingsConfiguration>();
-
         services.AddTransient<IConfigureOptions<ForwardedHeadersOptions>, ForwardedHeadersOptionsConfiguration>();
     }
 }
